+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firValueParameter.kt
@@ -1,24 +1,59 @@
 package com.javiersc.kotlin.compiler.extensions.fir
 
+import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.declarations.FirContextReceiver
 import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirValueParameter
 import org.jetbrains.kotlin.fir.declarations.builder.FirValueParameterBuilder
+import org.jetbrains.kotlin.fir.declarations.builder.buildRegularClass
 import org.jetbrains.kotlin.fir.declarations.builder.buildValueParameter
-import org.jetbrains.kotlin.fir.expressions.FirExpression
+import org.jetbrains.kotlin.fir.declarations.impl.FirResolvedDeclarationStatusImpl.Companion.DEFAULT_STATUS_FOR_STATUSLESS_DECLARATIONS
 import org.jetbrains.kotlin.fir.moduleData
+import org.jetbrains.kotlin.fir.scopes.kotlinScopeProvider
 import org.jetbrains.kotlin.fir.symbols.impl.FirClassLikeSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirClassSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirFunctionSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirRegularClassSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirValueParameterSymbol
-import org.jetbrains.kotlin.fir.types.builder.buildResolvedTypeRef
+import org.jetbrains.kotlin.fir.types.FirTypeRef
+import org.jetbrains.kotlin.fir.types.coneType
+import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.Name
 
-public fun FirClassLikeSymbol<*>.toValueParameter(
+public fun FirTypeRef.toValueParameter(
     session: FirSession,
     origin: FirDeclarationOrigin,
     containingFunctionSymbol: FirFunctionSymbol<*>,
-    defaultValue: FirExpression? = null,
+    block: FirValueParameterBuilder.(FirTypeRef) -> Unit = {},
+): FirValueParameter {
+    val typeRef: FirTypeRef = this@toValueParameter
+    val typeClassId: ClassId = typeRef.toClassId(session)
+    val symbol = FirRegularClassSymbol(typeClassId)
+    val typeParameters = typeRef.coneType.typeArguments
+    val firTypeParameters = typeParameters.map { it.toFirTypeParameter(session, origin, symbol) }
+    val classSymbol: FirRegularClassSymbol =
+        buildRegularClass {
+                this.moduleData = session.moduleData
+                this.origin = origin
+                this.typeParameters.addAll(firTypeParameters)
+                this.status = DEFAULT_STATUS_FOR_STATUSLESS_DECLARATIONS
+                this.classKind = ClassKind.CLASS
+                this.scopeProvider = session.kotlinScopeProvider
+                this.name = typeRef.toClassId(session).shortClassName
+                this.symbol = symbol
+            }
+            .symbol
+
+    return classSymbol.toValueParameter(session, origin, containingFunctionSymbol) {
+        block(this, typeRef)
+    }
+}
+
+public fun FirClassSymbol<*>.toValueParameter(
+    session: FirSession,
+    origin: FirDeclarationOrigin,
+    containingFunctionSymbol: FirFunctionSymbol<*>,
     block: FirValueParameterBuilder.(FirClassLikeSymbol<*>) -> Unit = {},
 ): FirValueParameter {
     val classSymbol: FirClassLikeSymbol<*> = this@toValueParameter
@@ -26,11 +61,10 @@ public fun FirClassLikeSymbol<*>.toValueParameter(
     return buildValueParameter {
         this.moduleData = session.moduleData
         this.origin = origin
-        this.returnTypeRef = buildResolvedTypeRef { this.type = classSymbol.classId.toConeType() }
+        this.returnTypeRef = classSymbol.toFirTypeRef()
         this.name = name
         this.symbol = FirValueParameterSymbol(name)
         this.containingFunctionSymbol = containingFunctionSymbol
-        this.defaultValue = defaultValue
         this.isCrossinline = false
         this.isNoinline = false
         this.isVararg = false